requirements.txt
capstone
pwntools
